requirements.txt
gradio>=4.0.0
numpy>=1.21.0
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0
sounddevice>=0.4.6
python-dotenv>=1.0.0
telethon>=1.31.1 